Start implementing a proper constructor for the gifti class
This is to make it compatible with MATLAB's namespaces.

Added properties with beginnings of MATLAB property validation such that
the code generates an empty gifti object when the constructor is given
no arguments. However, it is unclear how to test this properly.

Ispecially interesting is the struct branch of the constructor, where
the given struct contains the fields 'metadata', 'label' and 'data'.
It seems that the original behaviour would have resulted in infinite
recursion (without testing). Now the object is just returned as-is.

You will want to disable whitespace diffs when reviewing this one.